Sign on SUSE's Open Build Service
OBS can do the signing at build time, in 2 passes. The detached
hash(es) of the EFI files, taken with pesign, need to be copied into
../OTHER during the first pass, and named $PKG.cpio.rsasign which
will trigger the build script to send the cpio archive for signing
to the sign service, and then deposit the signed cpio archive in
../SOURCES and trigger a new build.
The second pass can then attach the signature to the EFI binary(-ies)
and repackage them.